Revert "Added finance GUI"
