--- Main Folder for GUI/FinancialGUI.java
@@ -1,4 +1,0 @@
-
-public class FinancialGUI {
-
-}
